+++ professional_templates/models/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 #import controllers
 import res_company
+import res_currency
 import invoice
 import sale_order
 import purchase_order